requirements.txt
-i https://pypi.org/simple
dnspython==2.0.0
pyyaml==5.3.1
